fix: initialState moving
