+++ src/components/Users/UsersContainer.js
@@ -5,19 +5,19 @@ import { followActionCreator, unfollowActionCreator, setUsersActionCreator } fro
 import Users from './Users';
 
 const mapStateToProps = (state) => ({
-    users: state.usersPage.users,
+  users: state.usersPage.users,
 });
 
 const mapDispatchToProps = (dispatch) => ({
-    follow: (userID) => {
-        dispatch(followActionCreator(userID));
-    },
-    unfollow: (userID) => {
-        dispatch(unfollowActionCreator(userID));
-    },
-    setUsers: (users) => {
-        dispatch(setUsersActionCreator(users));
-    },
+  follow: (userID) => {
+    dispatch(followActionCreator(userID));
+  },
+  unfollow: (userID) => {
+    dispatch(unfollowActionCreator(userID));
+  },
+  setUsers: (users) => {
+    dispatch(setUsersActionCreator(users));
+  },
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(Users);
